requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
strawberry-graphql==0.211.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
